feature(stab-fee): move poolpairs into utils dir, remove scan url mapping
